Bug 1801028: Improve dashboard single-stat number formatting
* Share `formatValue` utility between single state and table cards
* Honor font-size from single state panel definition

https://bugzilla.redhat.com/show_bug.cgi?id=1801028
